+++ CHANGELOG.md
@@ -14,8 +14,9 @@ All notable changes qsf-integration project will be documented in this file.
 - update commons-codec version
 - update slf4j-simple version
 - update maven version
-- ElasticWriter - check if the index exists before creating a new one
-- ElasticWriter - implement delete for bulk methods
+- improve thread handling in pipeline executer
+- improve thread handling in Solr Filter
+- improve handling of solr clients
 
 ### Fixed